backend/requirements.txt
websockets>=12.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-cloud-core>=2.3.0
requests>=2.31.0